Move display endpoint to multichain/network endpoint
